register mpfr_free_cache() callback only once per place
Registering a custodian callback once for each instantiation
causes an accumulation of callbacks when the custodian is not
explicitly shutdown.

The callbacks tend to close over a lot of data, despite efforts in the
Racket runtime system to limit the damage, probably because some
pieces of the callback are delay-loaded. Possibly, the runtime system
could be improved further, but it's better to not accumulate
callbacks, anyway.