use `_pointer` instead of `_gcpointer` in `_mpfr` record
On Racket CS, `_gcpointer` is a different representation (still
tagged) on read from a foreign struct. In this case, the pointer is a
self-reference to an immobile object, so `_pointer` is ok.

Using `_gcpointer` turns out to be harmless, even on CS, because we
don't read the field. But it can be confusing if you try to read the
field when debugging.