modular-expt: performance improvements
First, if `b` > 0, then in the loop, `b` will never reach 0, so we can
lift the case where `b` = 0 up over the loop. The answer in this case is
either 0 or 1, depending on the value of `n`.

Second, we can compute `(modulo a n)` just once. This has two benefits.
The obvious one is to avoid computing the potentially expensive
`(modulo a n)` (when `a` is huge) multiple times. The second one is to
avoid computing large multiplication `(* a ...)` in the odd
exponent case.

For example:

```
(define N 100000)

(define a (expt 2 N))
(define b (expt 3 N))
(define n 1000003)

(time (modular-expt a b n))
```

currently outputs:

```
cpu time: 5816 real time: 5816 gc time: 54
```

With the PR, it outputs:

```
cpu time: 3446 real time: 3446 gc time: 58
```