Correct license information.
The code in these packages was intended to be relicensed in 2019
with the rest of Racket (see racket/racket#1570). However, these
files were not changed due to an oversight.

Also, mention the dependence of "math-lib" on two LGPL libraries
which are distributed separately.

Also, correct the discussion about the license of some old Fortran
code.